Split renewers into individual modules, disable `dlink` by default
This commit finally separates renewers from the main `renewer.rs` file
and splits them into modules. This also means that renewers can now be
feature-gated easily. To reduce the number of dependencies for a basic
build of Xenon, the renewer `dlink` now needs to be enabled with the
feature flag `renewer-dlink`. If not enabled, the dependencies `sha2`,
`hmac` and `http` are not required anymore.

To compile Xenon with support for the `dlink` renewer, run:

cargo --build --features renewer-dlink

Plus any other feature you'd like to enable.