Finalize error handling: migrate from local errors to `error-chain`
This commit checks another mark on my TODO list for this project, and
makes oxixenon ready for the 1.0 release.

Oxixenon is now using `error-chain` for all of its errors, allowing it
to propagate errors and be as explicit as possible when reporting them.
This is a substantial architecture change and basically every file
in the project was involved by it.

Perks included with this commit:
- error messages shown in logs now also show their original cause
- clients only get the first error in the stack, which is the one with
  fewer details - this protects any kind of leak of information through
  errors.
- much less code related to error boilerplate and so on

This commit also ensures that when enabling client or server specific
features, the client and server features themselves are actually
included.